+++ src/App.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import {BrowserRouter, Routes, Route} from 'react-router-dom'
-import Comingsoon2 from './pages/Comingsoon2'
+import Comingsoon from './pages/Comingsoon'
 import NewPage from './components/NewPage'
-import Greet from './pages/Greet'
+import Greet from './pages/greet/Greet'
 const App = () => {
   return (
     <div>
       <BrowserRouter>
         <Routes>
-          <Route path='/' element={<Comingsoon2 />} />
+          <Route path='/' element={<Comingsoon />} />
           <Route path='/newpage' element={<NewPage />} />
           <Route path='/greet' element={<Greet />} />
         </Routes>
