--- src/components/Notifyme/NotifyMe.css
@@ -1,3 +1,0 @@
-.dialogBox{
-    margin: 15px;
-}
